+Completata la gestione delle stelle per Locale Business e Visita Business e il recupero di esse dal database

+++ PlaceHub_Java/src/database/BusinessDAO.java
@@ -24,7 +24,7 @@ public class BusinessDAO {
 		ResultSet datiRecuperati = query.executeQuery();
 		
 		while(datiRecuperati.next())
-			locali.add(new Locale(datiRecuperati.getString(1), datiRecuperati.getString(2), datiRecuperati.getString(3), datiRecuperati.getFloat(4),
+			locali.add(new Locale(datiRecuperati.getString(1), datiRecuperati.getString(2), datiRecuperati.getString(3), datiRecuperati.getDouble(4),
 					datiRecuperati.getString(5)));
 
 		return locali;
